+++ heksher/__init__.py
@@ -1,6 +1,6 @@
-from heksher.setting import Setting
-from heksher.clients.subclasses import TRACK_ALL
 from heksher.clients.async_client import AsyncHeksherClient
+from heksher.clients.subclasses import TRACK_ALL
 from heksher.clients.thread_client import ThreadHeksherClient
+from heksher.setting import Setting
 
-__all__ = ['Setting', 'AsyncHeksherClient', 'ThreadHeksherClient', 'TRACK_ALL']
+__all__ = ['AsyncHeksherClient', 'TRACK_ALL', 'ThreadHeksherClient', 'Setting']