+++ scripts/lint.sh
@@ -1,14 +1,10 @@
 #!/bin/sh
 # run various linters
 set -e
-poetry run flake8 --max-line-length 120 heksher
-set +e
-poetry run python -c "import pytype"
-res=$?
-set -e
-if [ "$res" -ne "0" ]
-  then
-    echo "pytype not run, please run in python 3.8 or lower"
-  else
-    poetry run pytype --keep-going heksher
-fi
+echo "running isort..."
+python -m isort . -c
+echo "running flake8..."
+python -m flake8 --max-line-length 120 heksher
+python -m flake8 --max-line-length 120 --extend-ignore=F841 tests
+echo "running mypy..."
+python3 -m mypy --show-error-codes heksher